% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{add_shockinfo}
\alias{add_shockinfo}
\title{Adding shocks}
\usage{
add_shockinfo(shockinfo=NULL, shock=NULL, restrictions=NULL, sign=NULL, horizon=NULL, scale=NULL)
}
\arguments{
\item{shockinfo}{Dataframe to append shocks. If \code{NULL} dataframe will be created.}

\item{shock}{Variable of structural shock. Only possible to add restrictions to one structural shock at a time.}

\item{restrictions}{Possible more than one variable which should be sign-restricted.}

\item{sign}{Possible more than one sign with which restrictions should be restricted.}

\item{horizon}{Possible more than one horizon with which restrictions should be restricted.}

\item{scale}{Scaling parameter, either positive (=1) or negative (=-1).}
}
\description{
Adding shocks
}
\details{
Function that helps to easily add shocks to 'shockinfo' collecting all sign restrictions.
}
