% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivar.R
\name{bivar}
\alias{bivar}
\title{Bayesian Interacted Vector Autoregression}
\usage{
bivar(Data, plag=1,draws=5000,burnin=5000,prior="NG",SV=TRUE,h=0,thin=1,
             hyperpara=NULL,eigen=FALSE,
             Ex=NULL,cons=FALSE,trend=FALSE,applyfun=NULL,cores=NULL,verbose=TRUE)
}
\arguments{
\item{Data}{Data in matrix form.}

\item{Idata}{Data for interaction term in matrix form.}

\item{plag}{number of lags}

\item{draws}{number of saved draws.}

\item{burnin}{number of burn-ins.}

\item{prior}{which prior}

\item{SV}{If set to \code{TRUE} stochastic volatility is enabled.}

\item{h}{holdout-sample.}

\item{thin}{thinning factor}

\item{hyperpara}{hyperparameter set}

\item{eigen}{should eigenvalues be computed?}

\item{Ex}{exogenous variables to add to the model}

\item{cons}{If set to \code{TRUE} a constant is included.}

\item{trend}{If set to \code{TRUE} a trend is included.}

\item{applyfun}{parallelization}

\item{cores}{number of cores}

\item{verbose}{verbosity option}
}
\description{
Bayesian Interacted Vector Autoregression
}
