% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTSM-package.R
\docType{package}
\name{BTSM-package}
\alias{BTSM}
\alias{BTSM-package}
\title{BTSM: Bayesian Time Series Models}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Maximilian Boeck \email{maximilian.boeck@wu.ac.at} (\href{https://orcid.org/0000-0001-6024-8305}{ORCID})

}
\keyword{internal}
