% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar.R
\name{bvar_natconj}
\alias{bvar_natconj}
\title{Bayesian Vector Autoregression with Natural Conjugate Prior setup}
\usage{
bvar(Yraw, plag, Wraw=NULL, draws=5000, cons=FALSE, h=0,
            sv=TRUE, shrink.type="lagwise", sample_theta=FALSE,
            cfit=FALSE, crit_eig=1.00, prmean=1)
}
\arguments{
\item{Yraw}{Data in matrix form}

\item{plag}{number of lags}

\item{Wraw}{exogenous variables.}

\item{draws}{number of saved draws.}

\item{cons}{If set to \code{TRUE} a constant is included.}

\item{h}{holdout-sample.}

\item{cfit}{whether log-likelihood should be computed}

\item{crit_eig}{critical eigenvalue}

\item{prmean}{prior mean.}

\item{sv}{If set to \code{TRUE} stochastic volatility is enabled.}

\item{shrink.type}{which kind of shrinkage should be applied. Either \code{global} or \code{lagwise}.}

\item{sample_theta}{whether \code{theta} should be sampled}
}
\description{
Bayesian Vector Autoregression with Natural Conjugate Prior setup
}
