% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{cond.predict}
\alias{cond.predict}
\title{Conditional Forecasts}
\usage{
cond.predict(constr, bvar.obj, pred.obj, constr_sd=NULL, verbose=TRUE)
}
\arguments{
\item{constr}{a matrix containing the conditional forecasts of size horizon times K, where horizon corresponds to the forecast horizon specified in \code{pred.obj}, while K is the number of variables in the system. The ordering of the variables have to correspond the ordering of the variables in the system. Rest is just set to NA.}

\item{bvar.obj}{an item fitted by \code{bvar}.}

\item{pred.obj}{an item fitted by \code{predict}. Note that \code{save.store=TRUE} is required as argument!}

\item{constr_sd}{a matrix containing the standard deviations around the conditional forecasts. Must have the same size as \code{constr}.}

\item{verbose}{If set to \code{FALSE} it suppresses printing messages to the console.}
}
\description{
Conditional Forecasts
}
\details{
Conditional forecasts need a fully identified system. Therefore this function utilizes short-run restrictions via the Cholesky decomposition on the global solution of the variance-covariance matrix of the Bayesian VAR.
}
