% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{get_shockinfo}
\alias{get_shockinfo}
\title{Create 'shockinfo' dataframe}
\usage{
get_shockinfo()
}
\description{
Create 'shockinfo' dataframe
}
\details{
Small helper function that creates a dataframe called 'shockinfo' which can be directly used to apply sign restrictions.
}
