% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd.R
\name{hd}
\alias{hd}
\alias{hd.bvar.irf}
\title{Historical Decomposition}
\usage{
hd(x, R=NULL, verbose=TRUE)
}
\arguments{
\item{x}{an item fitted by \code{IRF}.}

\item{R}{If \code{NULL} and the \code{irf.bgvar} object has been fitted via sign restrictions, the rotation matrix is used that minimizes the distance to the median impulse responses at the posterior median.}

\item{verbose}{If set to \code{FALSE} it suppresses printing messages to the console.}
}
\description{
A function that calculates historical decomposition (HD) of the time series and the structural error.
}
\details{
To save computational time as well as due to storage limits, both functions are based on the posterior median (as opposed to calculating HDs and the structural error for each draw of the MCMC chain). In case the shock has been identified via sign restrictions, a rotation matrix has to be selected to calculate both statistics. If not specified otherwise (via \code{R}), the algorithm searches for 50 rotation matrices that fulfill the sign restrictions at the \emph{posterior median} of the coefficients and then singles out the rotation matrix that minimizes the distance to the median of the impulse responses as suggested in Fry and Pagan (2011).
}
