% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivar.R, R/bpvar.R, R/bvar.R
\name{logLik}
\alias{logLik}
\alias{logLik.bivar}
\alias{logLik.bpvar}
\alias{logLik.bvar}
\title{Extract Log-likelihood of Bayesian VAR}
\usage{
\method{logLik}{bivar}(object, ..., quantile = 0.5)

\method{logLik}{bpvar}(object, ..., quantile = 0.5)

\method{logLik}{bvar}(object, ..., quantile = 0.5)
}
\arguments{
\item{object}{an object of class \code{bvar}.}

\item{...}{additional arguments.}

\item{quantile}{reported quantiles. Default is set to median.}
}
\value{
Returns an vector of dimension \code{q} (number of specified quantiles) of log-likelihoods.

Returns an vector of dimension \code{q} (number of specified quantiles) of log-likelihoods.

Returns an vector of dimension \code{q} (number of specified quantiles) of log-likelihoods.
}
\description{
Extract Log-Likelihood for \code{bivar}.

Extract Log-Likelihood for \code{bvar}.

Extract Log-Likelihood for \code{bvar}.
}
