% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{lps}
\alias{lps}
\alias{lps.bvar.pred}
\title{Compute Log-predictive Scores}
\usage{
\method{lps}{bvar.pred}(object, ...)
}
\arguments{
\item{object}{an object of class \code{bvar.predict}.}

\item{...}{additional arguments.}
}
\value{
Returns an object of class \code{bvar.lps}, which is a matrix of dimension h times K, whereas h is the forecasting horizon and K is the number of variables in the system.
}
\description{
Computes and prints log-predictive score of an object of class \code{bvar.predict}.
}
\author{
Maximilian Boeck
}
