% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict.bvar}
\title{Predictions}
\usage{
\method{predict}{bvar}(object, ..., n.ahead = 1, save.store = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{an object of class \code{bvar}.}

\item{...}{additional arguments.}

\item{n.ahead}{the forecast horizon.}

\item{save.store}{If set to \code{TRUE} the full distribution is returned. Default is set to \code{FALSE} in order to save storage.}

\item{verbose}{If set to \code{FALSE} it suppresses printing messages to the console.}
}
\description{
Predictions
}
