% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{rmse}
\alias{rmse}
\alias{rmse.bvar.pred}
\title{Compute Root Mean Squared Errors}
\usage{
\method{rmse}{bvar.pred}(object, ...)
}
\arguments{
\item{object}{an object of class \code{bvar.predict}.}

\item{...}{additional arguments.}
}
\value{
Returns an object of class \code{bvar.rmse}, which is a matrix of dimension h times K, whereas h is the forecasting horizon and K is the number of variables in the system.
}
\description{
Computes and prints root mean squared errors (RMSEs) of an object of class \code{bvar.predict}.
}
\author{
Maximilian Boeck
}
