% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvpbvarsim.R
\name{tvpbvar.sim}
\alias{tvpbvar.sim}
\title{Simulating a Time-varying Parameter Bayesian Vector Autoregression}
\usage{
tvpbvar.sim(len, M, plag=1, cons=FALSE, trend=FALSE, SV=FALSE)
}
\arguments{
\item{len}{length of the simulated time series.}

\item{M}{number of endogenous variables.}

\item{plag}{number of lags.}

\item{cons}{logical indicating whether to include an intercept. Default set to \code{FALSE}.}

\item{trend}{logical indicating whether to include an intercept. Default set to \code{FALSE}.}

\item{SV}{logical indicating whether the process should be simulated with or without stochastic volatility. Default set to \code{FALSE}.}

\item{sparse.coef}{sparsification of coefficients, has to be provided as percentage between zero and one.}

\item{sparse.tvp}{sparsification of time-variation, has to be provided as percentage between zero and one.}

\item{tvp.var}{variance of state process governing the time-variation in the coefficients}
}
\value{
Returns a list with the following elements
}
\description{
This function is used to produce simulated realizations which follow a Vector Autorgression (GVAR). It will also automatically simulate coefficients. All parameters can also be set by the user.
}
\details{
For testing purposes, this function enables to simulate time series processes which can be described by a Global Vector Autoregression. Since stability conditions are not checked, it is only implemented for \code{M=3}.
}
\examples{
library(BTSM)
sim <- tvpbvar.sim(len=200, M=3, plag=1, cons=TRUE, trend=FALSE, SV=FALSE)
Data = sim$obs$xglobal
W    = sim$obs$W
}
\author{
Maximilian Boeck
}
