% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar.R
\name{bvar}
\alias{bvar}
\title{Bayesian Vector Autoregression}
\usage{
bvar(
  Yraw,
  Wraw = NULL,
  p,
  nsave = 5000,
  nburn = 5000,
  cons = FALSE,
  fcst = FALSE,
  fhorz = 0,
  h = 0,
  sv = TRUE,
  shrink.type = "lagwise",
  sample_theta = FALSE,
  cfit = FALSE,
  store.post = TRUE,
  crit_eig = 1,
  Multiplier = 5,
  prmean = 1
)
}
\description{
Bayesian Vector Autoregression
}
