% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvecsim.R
\name{bvec.sim}
\alias{bvec.sim}
\title{Simulating a Vector Error Correction Model}
\usage{
bvec.sim(len, M, B, beta, plag=1, include="none")
}
\arguments{
\item{len}{length of the simulated time series.}

\item{M}{number of endogenous variables.}

\item{B}{matrix of coefficients.}

\item{beta}{cointegrating vector.}

\item{plag}{number of lags.}

\item{include}{either "none", "const", "trend" or "both" for including deterministics.}
}
\value{
Returns a list with the following elements
}
\description{
This function is used to produce simulated realizations which follow a Vector error correction model. It will also automatically simulate coefficients. All parameters can also be set by the user.
}
\details{
For testing purposes, this function enables to simulate time series processes which can be described by a Global Vector Autoregression. Since stability conditions are not checked, it is only implemented for \code{M=3}.
}
\examples{
len <- 1000
M <- 3
plag <- 1
include <- "none"
B <- rbind(c(0.7, 0, 0, 0.4), c(0.6, -0.5, 0.2, 0), c(0.2, 0, 0.4, 0.9))
beta <- c(1,-1,-1)
yb <- bvec.sim(len=len, M=M, B=B, beta=beta, plag=plag, include=include)
}
\author{
Maximilian Boeck
}
