% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf}
\alias{irf}
\title{Impulse Response Function}
\usage{
irf(x, n.ahead=24, ident=NULL, shockinfo=NULL, save.store=FALSE,
   applyfun=NULL, cores=NULL, verbose=TRUE,
   quantiles=c(.05,.10,.16,.50,.84,.90,.95), ...)
}
\arguments{
\item{x}{object of class \code{bvar}.}

\item{n.ahead}{forecasting horizon.}

\item{ident}{preferred identification scheme.}

\item{shockinfo}{dataframe with specified details on shock}

\item{save.store}{If set to \code{TRUE} the full posterior is returned. Default is set to \code{FALSE} in order to save storage.}

\item{applyfun}{Allows for user-specific apply function, which has to have the same interface than \code{lapply}. If \code{cores=NULL} then \code{lapply} is used, if set to a numeric either \code{parallel::parLapply()} is used on Windows platforms and \code{parallel::mclapply()} on non-Windows platforms.}

\item{cores}{Specifies the number of cores which should be used. Default is set to \code{NULL} and \code{applyfun} is used.}

\item{verbose}{If set to \code{FALSE} it suppresses printing messages to the console.}
}
\description{
Impulse Response Function
}
